working_paper_version/requirements.txt
# Water & Food Infrastructure Monitoring Pipeline
# v1.0.0
#
# Install all dependencies:
#   pip install -r requirements.txt
#
# numpy must be pinned first — numpy 2.x is required by rasterio >= 1.3 and
# ultralytics >= 8.  If you see a "numpy.dtype size changed" ABI error, run:
#   pip install --force-reinstall --no-cache-dir numpy
#   pip install -r requirements.txt

# ── Core ──────────────────────────────────────────────────────────────────────
numpy>=2.0.0,<3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
shapely>=2.0.0
geopandas>=0.14.0
pyproj>=3.6.0
tqdm>=4.66.0
pyyaml>=6.0.1

# ── Satellite imagery processing ─────────────────────────────────────────────
rasterio>=1.3.9
Pillow>=10.3.0

# ── ML / object detection ────────────────────────────────────────────────────
ultralytics>=8.2.0

# ── Optional: GPU acceleration (uncomment if using CUDA) ─────────────────────
# torch>=2.2.0
# torchvision>=0.17.0

# ── Optional: additional visualization ───────────────────────────────────────
# matplotlib>=3.8.0
# folium>=0.16.0     # alternative map rendering

# ── Development / testing ─────────────────────────────────────────────────────
# pytest>=8.0.0
# black>=24.0.0
# ruff>=0.4.0
